zutsuPT/pt/requirements.txt
Flask==3.1.*
Pillow==11.*
requests==2.*
